requirements.txt
torchaudio
omegaconf
ipython
soundfile
pygame